finagent/requirements.txt
browser-use
python-dotenv
langgraph
playwright
openai
langchain-openai

